Refactor AI architecture for improved separation of concerns
- Move all strategic decision logic from aiLogic.ts to new aiStrategy.ts
- Simplify getAIMove() to focus on API delegation and rule compliance
- Add getValidCombinations() to gameLogic.ts for centralized rule validation
- Update documentation to reflect new 6-module AI architecture
- Maintain all existing functionality while improving code organization
